+++ Sources/Container/GenericPersistentContaner.swift
@@ -44,17 +44,6 @@ private struct Default {
 ///
 public typealias AsyncErrorHandlerBlock = (Error) -> Void
 
-///
-/// Default block used to log Async errors if the user does not supply one
-///
-internal /// @testable
-func defaultAsyncErrorHandlingBlock(tag tagName: String) -> AsyncErrorHandlerBlock {
-
-    return { (error: Error) -> Void in
-        logError(tagName) { "\(error)" }
-    }
-}
-
 public enum GenericPersistentContainerErrors: Error {
     case invalidStoreDescription(String)
 }
@@ -62,12 +51,26 @@ public enum GenericPersistentContainerErrors: Error {
 ///
 /// A persistent container that can be customized with specific ContextStrategy Context type.
 ///
-public class GenericPersistentContainer<Strategy: ContextStrategyType>: PersistentStack {
+public class GenericPersistentContainer<Strategy: ContextStrategyType>: NSObject, PersistentContainer {
+
+    ///
+    /// Creates and returns a URL to the default directory for the persistent stores.
+    ///
+    public static func defaultStoreLocation() -> URL {
+        return Coherence.defaultStoreLocation()
+    }
 
     ///
     /// The name of this instance of GenericPersistentContainer.
     ///
     public let name: String
+
+    ///
+    /// The model this instance was constructed with.
+    ///
+    public var managedObjectModel: NSManagedObjectModel {
+        return persistentStoreCoordinator.managedObjectModel
+    }
 
     ///
     /// Returns the `NSPersistentStoreCoordinate` instance that
@@ -93,7 +96,20 @@ public class GenericPersistentContainer<Strategy: ContextStrategyType>: Persiste
     ///
     /// - Note: This method and the returned `BackgroundContext` can be used on a background thread.  It can also be used on the main thread.
     ///
-    public func newBackgroundContext<T: BackgroundContext>() -> T {
+    public func newBackgroundContext() -> BackgroundContext {
+        logInfo(self.tag) { "Creating new background context..." }
+        defer {
+            logInfo(self.tag) { "Background context created." }
+        }
+        return contextStrategy.newBackgroundContext()
+    }
+
+    ///
+    /// Gets a new `BackgroundContext` that can be used for updating objects.
+    ///
+    /// - Note: This method and the returned `BackgroundContext` can be used on a background thread.  It can also be used on the main thread.
+    ///
+    public func newGenericBackgroundContext<T: BackgroundContext>() -> T {
         logInfo(self.tag) { "Creating new background context of type `\(String(describing: T.self))`..." }
         defer {
             logInfo(self.tag) { "Background context created." }
@@ -113,7 +129,7 @@ public class GenericPersistentContainer<Strategy: ContextStrategyType>: Persiste
     ///
     /// - Parameters:
     ///     - name:             The name of the model file in the bundle. The model will be located based on the name given.
-    ///     - asyncErrorBlock:  An error handling block which will be called when an asynchronous error occurs (e.g. during a save of the main context to the persistent stores).
+    ///     - asyncErrorBlock:  An optional error handling block which will be called when an asynchronous error occurs (e.g. during a save of the contexts to the persistent stores).
     ///     - logTag:           An optional String that will be used as the tag for logging (default is GenericPersistentContainer).  This is typically used if you are embedding GenericPersistentContainer in something else and you want to to log as your class.
     ///
     /// - Returns: A generic container initialized with the given name.
@@ -138,7 +154,7 @@ public class GenericPersistentContainer<Strategy: ContextStrategyType>: Persiste
     /// - Parameters:
     ///     - name:               The name of the model file in the bundle.
     ///     - managedObjectModel: A managed object model.
-    ///     - asyncErrorBlock:    An error handling block which will be called when an asynchronous error occurs (e.g. during a save of the main context to the persistent stores).
+    ///     - asyncErrorBlock:    An optional error handling block which will be called when an asynchronous error occurs (e.g. during a save of the contexts to the persistent stores).
     ///     - logTag:             An optional String that will be used as the tag for logging (default is GenericPersistentContainer).  This is typically used if you are embedding GenericPersistentContainer in something else and you want to to log as your class.
     ///
     /// - Returns: A generic container initialized with the given name and model.
@@ -148,12 +164,29 @@ public class GenericPersistentContainer<Strategy: ContextStrategyType>: Persiste
         self.name = name
         self.tag  = logTag
 
-        self.errorHandlerBlock = asyncErrorBlock ?? defaultAsyncErrorHandlingBlock(tag: logTag)
+        self.errorHandlerBlock = { (error: Error) -> Void in
+
+            /// Log the error
+            logError(logTag) { "\(error)" }
+
+            /// Call the users block if supplied
+            if let errorBlock = asyncErrorBlock {
+                DispatchQueue.global().async {  /// Note: we move all user callback outside of our threads so that we can't be blocked by user code.
+                    errorBlock(error)
+                }
+            }
+        }
 
         /// Create the coordinator
         self.persistentStoreCoordinator = NSPersistentStoreCoordinator(managedObjectModel: model)
 
         self.contextStrategy = Strategy(persistentStoreCoordinator: self.persistentStoreCoordinator, errorHandler: self.errorHandlerBlock)
+    }
+
+    deinit {
+        do {
+            try self.detach(persistentStores: self.persistentStoreCoordinator.persistentStores)
+        } catch {/* ignore errors here */}
     }
 
     ///