+++ Tests/ContainerTests/GenericPersistentContainerTests.swift
@@ -102,20 +102,6 @@ class GenericPersistentContainerTests: XCTestCase {
         }
     }
 
-    func testDefaultErrorHandlingBlock() {
-
-        enum TestErrors: Error {
-            case testError
-        }
-
-        let input = TestErrors.testError
-
-        ///
-        /// Note: there is really no way at the moment to validate the output of the default handler so this is just to exercise it for test coverage completion.
-        ///
-        Coherence.defaultAsyncErrorHandlingBlock(tag: "Testing")(input)
-    }
-
     func testAttachPersistentStoresWithConfigurations() throws {
         
         let input = (modelName: "ContainerTestModel1", model: TestModelLoader.load(name: "ContainerTestModel1"), configuration: StoreConfiguration(fileName: "ContainerTestModel1.\(NSSQLiteStoreType.lowercased())"))